+++ R/rss.R
@@ -52,9 +52,9 @@ parse_feed <- function(doc) {
 #' \dontrun{
 #' crawler() |>
 #'   cr_on_html(\(ctx) ctx$push_data(list(
-#'     url = ctx$request$url, titulo = ctx$request$user_data$title
+#'     url = ctx$request$url, title = ctx$request$user_data$title
 #'   ))) |>
-#'   cr_from_rss("https://www.example.gov/noticias/rss")
+#'   cr_from_rss("https://example.com/feed.xml")
 #' }
 cr_from_rss <- function(crawler, url, label = NULL, include = NULL,
                         exclude = NULL, max = Inf) {